CREATE Utility function directory for reusable functions. AND CREATE Seller's page.
